requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pypdf
python-dotenv
faiss-cpu